services/ai-gateway/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pyyaml==6.0.2
google-cloud-storage==2.18.2
